Re-added loop for Windows
This is the only one common behaviour I got on local machine with Windows
and Travis CI Windows build.

So basically it looks like windows is locking file during reading,
so if we wait and try again it will be successful after reading.

The problem is that on some other Windows instances the file is also
locked during rename and it is not possible to read it (using include).
It is not a case on non-Windows env.

Added Travis CI configuration with Windows build - PHP 7.3

Signed-off-by: Michał Bundyra <contact@webimpress.com>